refac: add controllers dir #31
